rc_cloud/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
alembic==1.13.3
psycopg2-binary==2.9.9
pydantic==2.9.2
gspread==6.1.2
google-auth==2.35.0
python-dotenv==1.0.1

